Fixed mem leaks; unit vibration

+++ TestGameEnginev13/GameEngine2/unit.cpp
@@ -792,7 +792,7 @@ void unit::movement(bool siegesent) //the first time I put a class function outs
 	actspeed+=tilespeed[map[(unsigned int)y][(unsigned int)x].tilestyle-1]; //adjust speed for what tile the unit is standing on
 	if(sqrt(pow((movetox-x), 2)+pow((movetoy-y), 2))>actspeed+0.05)
 	{
-		if(abs((int)(movetox-x))>=0.0001)//if the line isn't vert
+		if(abs((double)(movetox-x))>=0.0001)//if the line isn't vert
 		{
 			float slope=(movetoy-y)/(movetox-x); //slope
 			float b=y-(slope*x); //y-intercept
